# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "KingMakers Mongo API",
    "version": "1.2.0",
    "description": "API documentation featuring OTP verification, Referral Leaderboards, and Multi-field Payment Updates."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/user/signup": {
      "post": {
        "summary": "Register a new user",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserInput" }
            }
          }
        },
        "responses": {
          "201": { "description": "User created successfully" },
          "400": { "description": "Phone number already exists" }
        }
      }
    },
    "/user/login": {
      "post": {
        "summary": "Login via plain-text password comparison",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phnumber": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["phnumber", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful" },
          "400": { "description": "Invalid password or user not found" }
        }
      }
    },
    "/api/verify-otp": {
      "post": {
        "summary": "Verify phone via Phone.email automated URL",
        "tags": ["Verification"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_json_url": { 
                    "type": "string", 
                    "example": "https://auth.phone.email/getuser/..." 
                  }
                },
                "required": ["user_json_url"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Phone verified" },
          "500": { "description": "Parsing error or request failed" }
        }
      }
    },
    "/user/updatepayment": {
      "post": {
        "summary": "Update payment using phnumber OR upi_id",
        "tags": ["User Actions"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phnumber": { "type": "string" },
                  "upi_id": { "type": "string" },
                  "investedamount": { "type": "number" }
                },
                "required": ["investedamount"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Payment updated" },
          "400": { "description": "Missing identifiers" },
          "404": { "description": "User not found" }
        }
      }
    },
    "/referral-leaderboard": {
      "get": {
        "summary": "Get leaderboard sorted by referral count",
        "tags": ["Referrals"],
        "responses": {
          "200": { "description": "Success" }
        }
      }
    },
    "/allusers": {
      "get": {
        "summary": "List all users",
        "tags": ["Admin"],
        "responses": {
          "200": { "description": "Success" }
        }
      }
    },
    "/user/delete": {
      "delete": {
        "summary": "Delete user by phone number",
        "tags": ["User Actions"],
        "parameters": [
          {
            "name": "phnumber",
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Deleted" },
          "404": { "description": "Not found" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UserInput": {
        "type": "object",
        "properties": {
          "phnumber": { "type": "string" },
          "name": { "type": "string" },
          "password": { "type": "string" },
          "dateofbirth": { "type": "string" },
          "gender": { "type": "string" },
          "occupation": { "type": "string" },
          "upi_id": { "type": "string" },
          "given_referral_code": { "type": "string" }
        },
        "required": ["phnumber", "name", "password", "dateofbirth", "gender", "occupation", "upi_id"]
      }
    }
  }
}